backend/requirements.txt
fastapi
uvicorn[standard]==0.32.0
python-multipart==0.0.9
Pillow==11.0.0
ultralytics
requests
python-dotenv